// src/api/__generated__/wallet/model/usdcIncomeLedgerItem.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama User Wallet API
 * 用户、钱包、任务中心接口
 * OpenAPI spec version: v1.0
 */
import type { UsdcIncomeLedgerItemBizType } from './usdcIncomeLedgerItemBizType';

/**
 * USDC 收入流水单项
 */
export interface UsdcIncomeLedgerItem {
  /** 交易时间戳（毫秒） */
  time?: number;
  /** 业务类型编码，枚举值：2101-签约费、2103-人工加款、2104-人工扣款、2106-体力费、2107-升级费、2108-手续费 */
  bizType?: UsdcIncomeLedgerItemBizType;
  /** 交易金额（USDC，精度 6 位小数，显示 4 位） */
  amount?: number;
}